trying to break up the functionality a bit
