+++ package.json
@@ -31,10 +31,9 @@
   },
   "homepage": "https://github.com/canjs/can-observe",
   "dependencies": {
-    "can-cid": "^1.0.1",
-    "can-key-tree": "0.0.3",
+    "can-key-tree": "<2.0.0",
     "can-namespace": "^1.0.0",
-    "can-observation-recorder": "^0.1.0",
+    "can-observation-recorder": "<2.0.0",
     "can-queues": "^0.2.5",
     "can-reflect": "^1.4.2",
     "can-symbol": "^1.2.0"
