+++ src/-symbols.js
@@ -1,0 +1,6 @@
+var canSymbol = require("can-symbol");
+
+module.exports = {
+    metaSymbol:canSymbol.for("can.meta"),
+    patchesSymbol: canSymbol("patches")
+};
