+++ test.html
@@ -1,2 +1,2 @@
 <title>can-observe tests</title>
-<script src="./node_modules/steal/steal.js" main="can-observe/test"></script>
+<script src="./node_modules/steal/steal.js" main="can-observe/test/test"></script>
