+++ test/test.js
@@ -1,0 +1,2 @@
+require("./object-test");
+require("./array-test");
